+++ src/pages/About/components/Sponsors/Sponsors.js
