+++ src/pages/About/components/Sponsors/Sponsors.scss
