+++ src/pages/About/components/Team/Team.scss
