+++ src/pages/Home/Home.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import './Home.scss';
+
+import Header from './components/Header/Header';
+import Bio from './components/Bio/Bio';
+import FAQ from './components/FAQ/FAQ';
+import Sponsors from '../About/components/Sponsors/Sponsors';
+import Gallery from './components/Gallery/Gallery';
+
+export default function Home() {
+   return (
+      <div className="home-container">
+        <Header />
+        <Bio/>
+        <Gallery />
+        <FAQ />
+        <Sponsors />
+      </div>
+    );
+}
