+++ src/pages/Home/components/Bio/Bio.scss
