+++ src/pages/Home/components/Events/Events.scss
