+++ src/pages/Home/components/FAQ/FAQ.scss
