+++ src/pages/Home/components/Footer/Footer.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import './Footer.scss';
-import Fishe from '../../assets/cfish.svg';
-import email from '../../assets/email.svg';
-import facebook from '../../assets/facebook.svg';
-import instagram from '../../assets/instagram.svg';
-import linkedin from '../../assets/linkedin.svg';
-import petal from '../../assets/GroupPetals.svg';
+import Fishe from '../../../../assets/cfish.svg';
+import email from '../../../../assets/email.svg';
+import facebook from '../../../../assets/facebook.svg';
+import instagram from '../../../../assets/instagram.svg';
+import linkedin from '../../../../assets/linkedin.svg';
+import petal from '../../../../assets/GroupPetals.svg';
 
 // import { makeStyles } from '@material-ui/core/styles';
