+++ src/pages/Home/components/Gallery/Gallery.scss
