+++ src/pages/Home/components/Header/Header.scss
@@ -2,11 +2,11 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
-  margin: 100px 0 20px;
+  margin: 0 0 20px;
   font-family: 'Open Sans', sans-serif;
 
   .banner {
-    background-image: url('../../assets/True2.svg');
+    background-image: url('../../../../assets/True2.svg');
     height: 500px;
     justify-content: center;
     align-items: center;
