+++ src/pages/Home/components/NavigationBar/NavigationBar.scss
